+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to specify how to match values together ([#15](https://github.com/khalyomede/reorder-before-after/issues/15)).
+
 ## [0.2.0] - 2023-01-14
 
 ### Added